Validate the name of the topic when it's being unsubscribed
Before this patch, when the bridge received the operation code 'unsubscribe',
it will not validate the topic name(e.g. validity of the topic or if it existed).

This patch implements that when receiving the 'unsubscribe' op, the bridge will
valide the name of the topic first. If the name itself is invalid, an error will
return. If the topic name doesn't exist, a warning will return.

Fix #87