Add client connection mode using the unused --address CLI argument (#132)
This PR allows the web bridge to be a websocket client that connects to the external server - requests can be issued by the server and proxied into the local network the same way a regular client would connect to the ros2-web-bridge as before. 

The --address arg is used to pass the address of the external server.